requirements.txt
aiohttp==3.8.1
#python-socketio==5.5.1
#python-engineio==4.3.1
Flask-SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0